pyproject.toml
[project]
name = "dnd-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.124.2",
    "httpx>=0.28.1",
    "mangum>=0.19.0",
    "pydantic-settings>=2.12.0",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.45",
    "uvicorn>=0.38.0",
]

[tool.pytest.ini_options]
pythonpath = "."
